+++ src/CfdiUtils/Elements/ImpLocal10/RetencionesLocales.php
@@ -1,0 +1,12 @@
+<?php
+namespace CfdiUtils\Elements\ImpLocal10;
+
+use CfdiUtils\Elements\Common\AbstractElement;
+
+class RetencionesLocales extends AbstractElement
+{
+    public function getElementName(): string
+    {
+        return 'implocal:RetencionesLocales';
+    }
+}